server/requirements.txt
fastapi~=0.115.5
pydantic~=2.9.2
starlette~=0.41.2
PyJWT~=2.9.0
uvicorn~=0.32.0
python-dotenv~=1.0.1